requirements.txt
# X-Platform Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
pyyaml>=6.0
pydantic>=2.0
numpy>=1.20.0
pandas>=1.3.0

# Vision dependencies
opencv-python>=4.5.0

# Protobuf (required by farm-ng SDK)
protobuf>=3.20.0

# farm-ng SDK (should already be installed)
# If not, install from: ~/Amiga/amiga-adk
# farm-ng-amiga>=2.0.0
# farm-ng-core>=2.0.0
